chore: move some tests over to test_round
