+++ py-polars/tests/unit/datatypes/test_temporal.py
@@ -2042,95 +2042,6 @@ def test_truncate_non_existent_14957() -> None:
         ).dt.truncate("46m")
 
 
-def test_round_ambiguous() -> None:
-    t = (
-        pl.datetime_range(
-            date(2020, 10, 25),
-            datetime(2020, 10, 25, 2),
-            "30m",
-            eager=True,
-            time_zone="Europe/London",
-        )
-        .alias("datetime")
-        .dt.offset_by("15m")
-    )
-    result = t.dt.round("30m")
-    expected = (
-        pl.Series(
-            [
-                "2020-10-25T00:30:00+0100",
-                "2020-10-25T01:00:00+0100",
-                "2020-10-25T01:30:00+0100",
-                "2020-10-25T01:00:00+0000",
-                "2020-10-25T01:30:00+0000",
-                "2020-10-25T02:00:00+0000",
-                "2020-10-25T02:30:00+0000",
-            ]
-        )
-        .str.to_datetime()
-        .dt.convert_time_zone("Europe/London")
-        .rename("datetime")
-    )
-    assert_series_equal(result, expected)
-
-    df = pl.DataFrame(
-        {
-            "date": pl.datetime_range(
-                date(2020, 10, 25),
-                datetime(2020, 10, 25, 2),
-                "30m",
-                eager=True,
-                time_zone="Europe/London",
-            ).dt.offset_by("15m")
-        }
-    )
-
-    df = df.select(pl.col("date").dt.round("30m"))
-    assert df.to_dict(as_series=False) == {
-        "date": [
-            datetime(2020, 10, 25, 0, 30, tzinfo=ZoneInfo("Europe/London")),
-            datetime(2020, 10, 25, 1, tzinfo=ZoneInfo("Europe/London")),
-            datetime(2020, 10, 25, 1, 30, tzinfo=ZoneInfo("Europe/London")),
-            datetime(2020, 10, 25, 1, tzinfo=ZoneInfo("Europe/London")),
-            datetime(2020, 10, 25, 1, 30, tzinfo=ZoneInfo("Europe/London")),
-            datetime(2020, 10, 25, 2, tzinfo=ZoneInfo("Europe/London")),
-            datetime(2020, 10, 25, 2, 30, tzinfo=ZoneInfo("Europe/London")),
-        ]
-    }
-
-
-def test_round_by_week() -> None:
-    df = pl.DataFrame(
-        {
-            "date": pl.Series(
-                [
-                    # Sunday and Monday
-                    "1998-04-12",
-                    "2022-11-28",
-                ]
-            ).str.strptime(pl.Date, "%Y-%m-%d")
-        }
-    )
-
-    assert (
-        df.select(
-            pl.col("date").dt.round("7d").alias("7d"),
-            pl.col("date").dt.round("1w").alias("1w"),
-        )
-    ).to_dict(as_series=False) == {
-        "7d": [date(1998, 4, 9), date(2022, 12, 1)],
-        "1w": [date(1998, 4, 13), date(2022, 11, 28)],
-    }
-
-
-@pytest.mark.parametrize("time_zone", [None, "Asia/Kathmandu"])
-def test_round_by_day_datetime(time_zone: str | None) -> None:
-    ser = pl.Series([datetime(2021, 11, 7, 3)]).dt.replace_time_zone(time_zone)
-    result = ser.dt.round("1d")
-    expected = pl.Series([datetime(2021, 11, 7)]).dt.replace_time_zone(time_zone)
-    assert_series_equal(result, expected)
-
-
 def test_cast_time_to_duration() -> None:
     assert pl.Series([time(hour=0, minute=0, second=2)]).cast(
         pl.Duration
